+++ TabletopTweaks-Base/Bugfixes/General/AreaOfEffectsTick.cs
@@ -22,7 +22,7 @@ namespace TabletopTweaks.Base.Bugfixes.General {
         static class AreaOfEffectsTick_Round_Patch {
             static void Postfix(AreaEffectEntityData __instance) {
                 if (Main.TTTContext.Fixes.BaseFixes.IsDisabled("AreaOfEffectDoubleTrigger")) { return; }
-                if (__instance.Blueprint.GetComponent<AbilityAreaEffectRunAction>().UnitEnter.HasActions) {
+                if (__instance.Blueprint.GetComponent<AbilityAreaEffectRunAction>()?.UnitEnter?.HasActions ?? false) {
                     __instance.m_TimeToNextRound = 6f;
                 }
             }